requirements.txt
streamlit==1.48.1
pandas==2.3.1
pydantic==2.11.7
uvicorn==0.35.0
mysql-connector-python==9.4.0
requests==2.32.4
fastapi==0.116.1